Clean up and organize tests a little.
Split pingpong tests and streaming tests into their own directories.
Return non-negated value from fi_getinfo for all tests when
failing to open device (61 FI_ENODATA).

Make command name be dictionary key for perf results.

Makes machine readable output look like:
---
./pingpong/fi_rdm_atomic -f sockets -m localhost :
- { xfer_size: 64, iterations: 100000, total: 12800000, time: 5.441129, Gb/sec: 0.018820, usec/xfer: 27.205645 }
- { xfer_size: 64, iterations: 100000, total: 12800000, time: 5.561684, Gb/sec: 0.018412, usec/xfer: 27.808420 }
- { xfer_size: 64, iterations: 100000, total: 12800000, time: 5.521032, Gb/sec: 0.018547, usec/xfer: 27.605160 }
