+++ Makefile.am
@@ -4,25 +4,25 @@ ACLOCAL_AMFLAGS = -I config
 bin_PROGRAMS = \
 	simple/fi_info \
 	simple/fi_msg\
-	simple/fi_msg_pingpong \
-	simple/fi_msg_rma \
 	simple/fi_rdm \
 	simple/fi_rdm_rma_simple \
 	simple/fi_dgram \
 	simple/fi_dgram_waitset \
-	simple/fi_rdm_pingpong \
-	simple/fi_rdm_tagged_pingpong \
 	simple/fi_rdm_tagged_search \
-	simple/fi_rdm_cntr_pingpong \
-	simple/fi_rdm_rma \
-	simple/fi_rdm_atomic \
-	simple/fi_ud_pingpong \
 	simple/fi_cq_data \
 	simple/fi_poll \
-	simple/fi_rdm_inject_pingpong \
-	simple/fi_rdm_multi_recv \
 	simple/fi_scalable_ep \
 	simple/fi_rdm_shared_ctx \
+	streaming/fi_msg_rma \
+	streaming/fi_rdm_atomic \
+	streaming/fi_rdm_multi_recv \
+	streaming/fi_rdm_rma \
+	pingpong/fi_msg_pingpong \
+	pingpong/fi_rdm_inject_pingpong \
+	pingpong/fi_rdm_cntr_pingpong \
+	pingpong/fi_ud_pingpong \
+	pingpong/fi_rdm_pingpong \
+	pingpong/fi_rdm_tagged_pingpong \
 	unit/fi_eq_test \
 	unit/fi_av_test \
 	unit/fi_size_left_test \
@@ -51,18 +51,6 @@ simple_fi_msg_SOURCES = \
 	simple/msg.c
 simple_fi_msg_LDADD = libfabtests.la
 
-simple_fi_msg_pingpong_SOURCES = \
-	simple/msg_pingpong.c
-simple_fi_msg_pingpong_LDADD = libfabtests.la
-
-simple_fi_msg_rma_SOURCES = \
-	simple/msg_rma.c
-simple_fi_msg_rma_LDADD = libfabtests.la
-
-simple_fi_msg_epoll_SOURCES = \
-	simple/msg_epoll.c
-simple_fi_msg_epoll_LDADD = libfabtests.la
-
 simple_fi_rdm_SOURCES = \
 	simple/rdm.c
 simple_fi_rdm_LDADD = libfabtests.la
@@ -79,41 +67,13 @@ simple_fi_dgram_waitset_SOURCES = \
 	simple/dgram_waitset.c
 simple_fi_dgram_waitset_LDADD = libfabtests.la
 
-simple_fi_rdm_pingpong_SOURCES = \
-	simple/rdm_pingpong.c
-simple_fi_rdm_pingpong_LDADD = libfabtests.la
-
-simple_fi_rdm_tagged_pingpong_SOURCES = \
-	simple/rdm_tagged_pingpong.c
-simple_fi_rdm_tagged_pingpong_LDADD = libfabtests.la
-
 simple_fi_rdm_tagged_search_SOURCES = \
 	simple/rdm_tagged_search.c
 simple_fi_rdm_tagged_search_LDADD = libfabtests.la
 
-simple_fi_rdm_cntr_pingpong_SOURCES = \
-	simple/rdm_cntr_pingpong.c
-simple_fi_rdm_cntr_pingpong_LDADD = libfabtests.la
-
-simple_fi_rdm_rma_SOURCES = \
-	simple/rdm_rma.c
-simple_fi_rdm_rma_LDADD = libfabtests.la
-
-simple_fi_rdm_atomic_SOURCES = \
-	simple/rdm_atomic.c
-simple_fi_rdm_atomic_LDADD = libfabtests.la
-
-simple_fi_ud_pingpong_SOURCES = \
-	simple/ud_pingpong.c
-simple_fi_ud_pingpong_LDADD = libfabtests.la
-
 simple_fi_cq_data_SOURCES = \
 	simple/cq_data.c
 simple_fi_cq_data_LDADD = libfabtests.la
-
-simple_fi_rdm_inject_pingpong_SOURCES = \
-	simple/rdm_inject_pingpong.c
-simple_fi_rdm_inject_pingpong_LDADD = libfabtests.la
 
 simple_fi_scalable_ep_SOURCES = \
 	simple/scalable_ep.c
@@ -127,9 +87,46 @@ simple_fi_poll_SOURCES = \
 	simple/poll.c
 simple_fi_poll_LDADD = libfabtests.la
 
-simple_fi_rdm_multi_recv_SOURCES = \
-	simple/rdm_multi_recv.c
-simple_fi_rdm_multi_recv_LDADD = libfabtests.la
+streaming_fi_msg_rma_SOURCES = \
+	streaming/msg_rma.c
+streaming_fi_msg_rma_LDADD = libfabtests.la
+
+streaming_fi_rdm_atomic_SOURCES = \
+	streaming/rdm_atomic.c
+streaming_fi_rdm_atomic_LDADD = libfabtests.la
+
+streaming_fi_rdm_multi_recv_SOURCES = \
+	streaming/rdm_multi_recv.c
+streaming_fi_rdm_multi_recv_LDADD = libfabtests.la
+
+streaming_fi_rdm_rma_SOURCES = \
+	streaming/rdm_rma.c
+streaming_fi_rdm_rma_LDADD = libfabtests.la
+
+pingpong_fi_msg_pingpong_SOURCES = \
+	pingpong/msg_pingpong.c
+pingpong_fi_msg_pingpong_LDADD = libfabtests.la
+
+pingpong_fi_ud_pingpong_SOURCES = \
+	pingpong/ud_pingpong.c
+pingpong_fi_ud_pingpong_LDADD = libfabtests.la
+
+pingpong_fi_rdm_cntr_pingpong_SOURCES = \
+	pingpong/rdm_cntr_pingpong.c
+pingpong_fi_rdm_cntr_pingpong_LDADD = libfabtests.la
+
+pingpong_fi_rdm_pingpong_SOURCES = \
+	pingpong/rdm_pingpong.c
+pingpong_fi_rdm_pingpong_LDADD = libfabtests.la
+
+pingpong_fi_rdm_tagged_pingpong_SOURCES = \
+	pingpong/rdm_tagged_pingpong.c
+pingpong_fi_rdm_tagged_pingpong_LDADD = libfabtests.la
+
+pingpong_fi_rdm_inject_pingpong_SOURCES = \
+	pingpong/rdm_inject_pingpong.c
+pingpong_fi_rdm_inject_pingpong_LDADD = libfabtests.la
+
 
 unit_fi_eq_test_SOURCES = \
 	unit/eq_test.c \
