+++ pingpong/rdm_inject_pingpong.c
@@ -1,12 +1,8 @@
 /*
  * Copyright (c) 2013-2014 Intel Corporation.  All rights reserved.
- * Copyright (c) 2014 Cisco Systems, Inc.  All rights reserved.
  *
- * This software is available to you under a choice of one of two
- * licenses.  You may choose to be licensed under the terms of the GNU
- * General Public License (GPL) Version 2, available from the file
- * COPYING in the main directory of this source tree, or the
- * BSD license below:
+ * This software is available to you under the BSD license
+ * below:
  *
  *     Redistribution and use in source and binary forms, with or
  *     without modification, are permitted provided that the following
@@ -38,7 +34,6 @@
 #include <time.h>
 #include <netdb.h>
 #include <unistd.h>
-#include <arpa/inet.h>
 
 #include <rdma/fabric.h>
 #include <rdma/fi_errno.h>
@@ -48,65 +43,36 @@
 #include "shared.h"
 
 static struct cs_opts opts;
+static int max_inject_size;
 static int max_credits = 128;
-static int credits = 128;
 static char test_name[10] = "custom";
 static struct timespec start, end;
-static void *buf;
-static void *buf_ptr;
+static void *send_buf, *recv_buf;
 static size_t buffer_size;
-static size_t prefix_len;
-static size_t max_msg_size = 0;
 
-static struct fi_info *hints;
-static fi_addr_t remote_fi_addr;
+static struct fi_info *fi, *hints;
 
-static struct fi_info *fi;
 static struct fid_fabric *fab;
 static struct fid_domain *dom;
 static struct fid_ep *ep;
-static struct fid_cq *rcq, *scq;
-static struct fid_mr *mr;
+static struct fid_cq *rcq;
+static struct fid_cq *scq;
 static struct fid_av *av;
-
-static int poll_all_sends(void)
-{
-	struct fi_cq_entry comp;
-	int ret;
-
-	do {
-		ret = fi_cq_read(scq, &comp, 1);
-		if (ret > 0) {
-			credits++;
-		} else if (ret < 0 && ret != -FI_EAGAIN) {
-			FT_PRINTERR("fi_cq_read", ret);
-			return ret;
-		}
-	} while (ret != -FI_EAGAIN);
-	return 0;
-}
+static struct fid_mr *mr;
+static void *local_addr, *remote_addr;
+static size_t addrlen = 0;
+static fi_addr_t remote_fi_addr;
+struct fi_context fi_ctx_send;
+struct fi_context fi_ctx_recv;
+struct fi_context fi_ctx_av;
 
 static int send_xfer(int size)
 {
-	struct fi_cq_entry comp;
 	int ret;
 
-	while (!credits) {
-		ret = fi_cq_read(scq, &comp, 1);
-		if (ret > 0) {
-			goto post;
-		} else if (ret < 0 && ret != -FI_EAGAIN) {
-			FT_PRINTERR("fi_cq_read", ret);
-			return ret;
-		}
-	}
-
-	credits--;
-post:
-	ret = fi_send(ep, buf_ptr, (size_t) size, fi_mr_desc(mr),
-			remote_fi_addr, NULL);
+	ret = fi_inject(ep, send_buf, (size_t) size, remote_fi_addr);
 	if (ret)
-		FT_PRINTERR("fi_send", ret);
+		FT_PRINTERR("fi_inject", ret);
 
 	return ret;
 }
@@ -117,16 +83,49 @@ static int recv_xfer(int size)
 	int ret;
 
 	do {
-		ret = fi_cq_read(rcq, &comp, sizeof comp);
+		ret = fi_cq_read(rcq, &comp, 1);
 		if (ret < 0 && ret != -FI_EAGAIN) {
-			FT_PRINTERR("fi_cq_read", ret);
+			if (ret == -FI_EAVAIL) {
+				cq_readerr(rcq, "rcq");
+			} else {
+				FT_PRINTERR("fi_cq_read", ret);
+			}
 			return ret;
 		}
 	} while (ret == -FI_EAGAIN);
 
-	ret = fi_recv(ep, buf, buffer_size, fi_mr_desc(mr), 0, buf);
+	ret = fi_recv(ep, recv_buf, buffer_size, fi_mr_desc(mr), remote_fi_addr,
+			&fi_ctx_recv);
 	if (ret)
 		FT_PRINTERR("fi_recv", ret);
+
+	return ret;
+}
+
+static int send_msg(int size)
+{
+	int ret;
+
+	ret = fi_inject(ep, send_buf, (size_t) size, remote_fi_addr);
+	if (ret) {
+		FT_PRINTERR("fi_inject", ret);
+		return ret;
+	}
+
+	return ret;
+}
+
+static int recv_msg(void)
+{
+	int ret;
+
+	ret = fi_recv(ep, recv_buf, buffer_size, fi_mr_desc(mr), 0, &fi_ctx_recv);
+	if (ret) {
+		FT_PRINTERR("fi_recv", ret);
+		return ret;
+	}
+
+	ret = wait_for_completion(rcq, 1);
 
 	return ret;
 }
@@ -134,9 +133,6 @@ static int recv_xfer(int size)
 static int sync_test(void)
 {
 	int ret;
-
-	while (credits < max_credits)
-		poll_all_sends();
 
 	ret = opts.dst_addr ? send_xfer(16) : recv_xfer(16);
 	if (ret)
@@ -149,21 +145,24 @@ static int run_test(void)
 {
 	int ret, i;
 
+	if (opts.transfer_size > max_inject_size) 
+		return 0;
+
 	ret = sync_test();
 	if (ret)
-		return ret;
+		goto out;
 
 	clock_gettime(CLOCK_MONOTONIC, &start);
 	for (i = 0; i < opts.iterations; i++) {
 		ret = opts.dst_addr ? send_xfer(opts.transfer_size) :
 				 recv_xfer(opts.transfer_size);
 		if (ret)
-			return ret;
+			goto out;
 
 		ret = opts.dst_addr ? recv_xfer(opts.transfer_size) :
 				 send_xfer(opts.transfer_size);
 		if (ret)
-			return ret;
+			goto out;
 	}
 	clock_gettime(CLOCK_MONOTONIC, &end);
 
@@ -172,38 +171,21 @@ static int run_test(void)
 	else
 		show_perf(test_name, opts.transfer_size, opts.iterations, &start, &end, 2);
 
-	return 0;
+	ret = 0;
+
+out:
+	return ret;
 }
 
 static void free_ep_res(void)
 {
-	int ret;
-
-	if (ep) {
-		ret = fi_close(&ep->fid);
-		if (ret != 0) {
-			FT_PRINTERR("fi_close", ret);
-		}
-	}
-	if (mr) {
-		ret = fi_close(&mr->fid);
-		if (ret != 0) {
-			FT_PRINTERR("fi_close", ret);
-		}
-	}
-	if (rcq) {
-		ret = fi_close(&rcq->fid);
-		if (ret != 0) {
-			FT_PRINTERR("fi_close", ret);
-		}
-	}
-	if (scq) {
-		ret = fi_close(&scq->fid);
-		if (ret != 0) {
-			FT_PRINTERR("fi_close", ret);
-		}
-	}
-	free(buf);
+	fi_close(&ep->fid);
+	fi_close(&av->fid);
+	fi_close(&mr->fid);
+	fi_close(&rcq->fid);
+	fi_close(&scq->fid);
+	free(send_buf);
+	free(recv_buf);
 }
 
 static int alloc_ep_res(struct fi_info *fi)
@@ -214,47 +196,44 @@ static int alloc_ep_res(struct fi_info *fi)
 
 	buffer_size = opts.user_options & FT_OPT_SIZE ?
 			opts.transfer_size : test_size[TEST_CNT - 1].size;
-	if (max_msg_size > 0 && buffer_size > max_msg_size) {
-		buffer_size = max_msg_size;
-	}
-	if (buffer_size < fi->src_addrlen) {
-		buffer_size = fi->src_addrlen;
-	}
-	buffer_size += prefix_len;
-	buf = malloc(buffer_size);
-	if (!buf) {
+	send_buf = malloc(buffer_size);
+	recv_buf = malloc(buffer_size);
+	if (!send_buf || !recv_buf) {
 		perror("malloc");
 		return -1;
 	}
-	buf_ptr = (char *)buf + prefix_len;
 
 	memset(&cq_attr, 0, sizeof cq_attr);
 	cq_attr.format = FI_CQ_FORMAT_CONTEXT;
 	cq_attr.wait_obj = FI_WAIT_NONE;
 	cq_attr.size = max_credits << 1;
-	ret = fi_cq_open(dom, &cq_attr, &scq, NULL);
+
+	ret = fi_cq_open(dom, &cq_attr, &rcq, NULL);
 	if (ret) {
 		FT_PRINTERR("fi_cq_open", ret);
 		goto err1;
 	}
 
-	ret = fi_cq_open(dom, &cq_attr, &rcq, NULL);
+	ret = fi_cq_open(dom, &cq_attr, &scq, NULL);
 	if (ret) {
 		FT_PRINTERR("fi_cq_open", ret);
 		goto err2;
 	}
 
-	ret = fi_mr_reg(dom, buf, buffer_size, 0, 0, 0, 0, &mr, NULL);
+	/* Memory registration not required for send_buf since we use fi_inject.
+	 * fi_inject copies the buffer of data that needs to be sent. */
+	ret = fi_mr_reg(dom, recv_buf, buffer_size, 0, 0, 0, 0, &mr, NULL);
 	if (ret) {
-		FT_PRINTERR("fi_mr_reg", ret);
+		FT_PRINTERR("fi_cq_open", ret);
 		goto err3;
 	}
 
-	memset(&av_attr, 0, sizeof(av_attr));
+	memset(&av_attr, 0, sizeof av_attr);
 	av_attr.type = fi->domain_attr->av_type ?
 			fi->domain_attr->av_type : FI_AV_MAP;
+	av_attr.count = 1;
 	av_attr.name = NULL;
-	av_attr.flags = 0;
+
 	ret = fi_av_open(dom, &av_attr, &av, NULL);
 	if (ret) {
 		FT_PRINTERR("fi_av_open", ret);
@@ -274,11 +253,12 @@ err5:
 err4:
 	fi_close(&mr->fid);
 err3:
-	fi_close(&rcq->fid);
-err2:
 	fi_close(&scq->fid);
+err2:
+	fi_close(&rcq->fid);
 err1:
-	free(buf);
+	free(send_buf);
+	free(recv_buf);
 	return ret;
 }
 
@@ -286,13 +266,13 @@ static int bind_ep_res(void)
 {
 	int ret;
 
-	ret = fi_ep_bind(ep, &scq->fid, FI_SEND);
+	ret = fi_ep_bind(ep, &rcq->fid, FI_RECV);
 	if (ret) {
 		FT_PRINTERR("fi_ep_bind", ret);
 		return ret;
 	}
 
-	ret = fi_ep_bind(ep, &rcq->fid, FI_RECV);
+	ret = fi_ep_bind(ep, &scq->fid, FI_SEND);
 	if (ret) {
 		FT_PRINTERR("fi_ep_bind", ret);
 		return ret;
@@ -310,21 +290,14 @@ static int bind_ep_res(void)
 		return ret;
 	}
 
-	/* Post the first recv buffer */
-	ret = fi_recv(ep, buf, buffer_size, fi_mr_desc(mr), 0, buf);
-	if (ret) {
-		FT_PRINTERR("fi_recv", ret);
-	}
-
 	return ret;
 }
 
-
-static int common_setup(void)
+static int init_fabric(void)
 {
-	int ret;
 	uint64_t flags = 0;
 	char *node, *service;
+	int ret;
 
 	if (opts.dst_addr) {
 		ret = ft_getsrcaddr(opts.src_addr, opts.src_port, hints);
@@ -341,146 +314,131 @@ static int common_setup(void)
 	ret = fi_getinfo(FT_FIVERSION, node, service, flags, hints, &fi);
 	if (ret) {
 		FT_PRINTERR("fi_getinfo", ret);
-		goto err1;
+		return ret;
 	}
-	if (fi->ep_attr->max_msg_size) {
-		max_msg_size = fi->ep_attr->max_msg_size;
+	
+	/* check max msg size */
+	max_inject_size = fi->tx_attr->inject_size;
+	if ((opts.user_options & FT_OPT_SIZE) &&
+	    (opts.transfer_size > max_inject_size)) {
+		fprintf(stderr, "Msg size greater than max inject size\n");
+		ret = -FI_EINVAL;
+		goto err0;
+	}
+
+	/* We use provider MR attributes and direct address (no offsets) 
+	 * for RMA calls */
+	if (!(fi->mode & FI_PROV_MR_ATTR))
+		fi->mode |= FI_PROV_MR_ATTR;
+
+	/* Get remote address */
+	if (opts.dst_addr) {
+		addrlen = fi->dest_addrlen;
+		remote_addr = malloc(addrlen);
+		memcpy(remote_addr, fi->dest_addr, addrlen);
 	}
 
 	ret = fi_fabric(fi->fabric_attr, &fab, NULL);
 	if (ret) {
 		FT_PRINTERR("fi_fabric", ret);
-		goto err1;
-	}
-	if (fi->mode & FI_MSG_PREFIX) {
-		prefix_len = fi->ep_attr->msg_prefix_size;
+		goto err0;
 	}
 
 	ret = fi_domain(fab, fi, &dom, NULL);
 	if (ret) {
 		FT_PRINTERR("fi_domain", ret);
-		goto err2;
+		goto err1;
 	}
 
 	ret = alloc_ep_res(fi);
-	if (ret) {
-		goto err4;
-	}
+	if (ret)
+		goto err3;
 
 	ret = bind_ep_res();
-	if (ret) {
-		goto err5;
-	}
+	if (ret)
+		goto err4;
 
-	if (hints->src_addr)
-		free(hints->src_addr);
 	return 0;
 
-err5:
-	free_ep_res();
 err4:
-	fi_close(&dom->fid);
-err2:
-	fi_close(&fab->fid);
-err1:
-	if (hints->src_addr)
-		free(hints->src_addr);
-	return ret;
-}
-
-static int client_connect(void)
-{
-	int ret;
-
-	ret = common_setup();
-	if (ret != 0)
-		goto err;
-
-	ret = ft_getdestaddr(opts.dst_addr, opts.dst_port, hints);
-	if (ret != 0)
-		goto err;
-
-	ret = fi_av_insert(av, hints->dest_addr, 1, &remote_fi_addr, 0, NULL);
-	if (ret != 1) {
-		FT_PRINTERR("fi_av_insert", ret);
-		goto err;
-	}
-
-	// send initial message to server with our local address
-	memcpy(buf_ptr, fi->src_addr, fi->src_addrlen);
-	ret = send_xfer(fi->src_addrlen);
-	if (ret != 0)
-		goto err;
-
-	// wait for reply to know server is ready
-	ret = recv_xfer(4);
-	if (ret != 0)
-		goto err;
-
-	return 0;
-
-err:
 	free_ep_res();
-	if (av) {
-		fi_close(&av->fid);
-	}
-	if (ep) {
-		fi_close(&ep->fid);
-	}
-	if (dom) {
-		fi_close(&dom->fid);
-	}
-	if (fab) {
-		fi_close(&fab->fid);
-	}
+err3:
+	fi_close(&dom->fid);
+err1:
+	fi_close(&fab->fid);
+err0:
 	return ret;
 }
 
-static int server_connect(void)
+static int init_av(void)
 {
 	int ret;
-	struct fi_cq_entry comp;
 
-	ret = common_setup();
-	if (ret != 0)
-		goto err;
-
-	do {
-		ret = fi_cq_read(rcq, &comp, 1);
-		if (ret < 0 && ret != -FI_EAGAIN) {
-			FT_PRINTERR("fi_cq_read", ret);
+	if (opts.dst_addr) {
+		/* Get local address blob. Find the addrlen first. We set addrlen 
+		 * as 0 and fi_getname will return the actual addrlen. */
+		addrlen = 0;
+		ret = fi_getname(&ep->fid, local_addr, &addrlen);
+		if (ret != -FI_ETOOSMALL) {
+			FT_PRINTERR("fi_getname", ret);
 			return ret;
 		}
-	} while (ret == -FI_EAGAIN);
 
-	ret = fi_av_insert(av, buf_ptr, 1, &remote_fi_addr, 0, NULL);
-	if (ret != 1) {
-		if (ret == 0) {
-			fprintf(stderr, "Unable to resolve remote address 0x%x 0x%x\n",
-				((uint32_t *)buf)[0], ((uint32_t *)buf)[1]);
-			ret = -FI_EINVAL;
-		} else {
-			FT_PRINTERR("fi_av_insert", ret);
+		local_addr = malloc(addrlen);
+		ret = fi_getname(&ep->fid, local_addr, &addrlen);
+		if (ret) {
+			FT_PRINTERR("fi_getname", ret);
+			return ret;
 		}
-		goto err;
+
+		ret = fi_av_insert(av, remote_addr, 1, &remote_fi_addr, 0, 
+				&fi_ctx_av);
+		if (ret != 1) {
+			FT_PRINTERR("fi_av_insert", ret);
+			return ret;
+		}
+
+		/* Send local addr size and local addr */
+		memcpy(send_buf, &addrlen, sizeof(size_t));
+		memcpy(send_buf + sizeof(size_t), local_addr, addrlen);
+		ret = send_msg(sizeof(size_t) + addrlen);
+		if (ret)
+			return ret;
+
+		/* Receive ACK from server */
+		ret = recv_msg();
+		if (ret)
+			return ret;
+
+	} else {
+		/* Post a recv to get the remote address */
+		ret = recv_msg();
+		if (ret)
+			return ret;
+
+		memcpy(&addrlen, recv_buf, sizeof(size_t));
+		remote_addr = malloc(addrlen);
+		memcpy(remote_addr, recv_buf + sizeof(size_t), addrlen);
+
+		ret = fi_av_insert(av, remote_addr, 1, &remote_fi_addr, 0, 
+				&fi_ctx_av);
+		if (ret != 1) {
+			FT_PRINTERR("fi_av_insert", ret);
+			return ret;
+		}
+
+		/* Send ACK */
+		ret = send_msg(16);
+		if (ret)
+			return ret;
 	}
 
-	ret = fi_recv(ep, buf, buffer_size, fi_mr_desc(mr), 0, buf);
-	if (ret != 0) {
+	/* Post first recv */
+	ret = fi_recv(ep, recv_buf, buffer_size, fi_mr_desc(mr), remote_fi_addr,
+			&fi_ctx_recv);
+	if (ret)
 		FT_PRINTERR("fi_recv", ret);
-		goto err;
-	}
 
-	ret = send_xfer(4);
-	if (ret != 0)
-		goto err;
-
-	return 0;
-
-err:
-	free_ep_res();
-	fi_close(&dom->fid);
-	fi_close(&fab->fid);
 	return ret;
 }
 
@@ -488,9 +446,13 @@ static int run(void)
 {
 	int i, ret = 0;
 
-	ret = opts.dst_addr ? client_connect() : server_connect();
+	ret = init_fabric();
 	if (ret)
 		return ret;
+
+	ret = init_av();
+	if (ret)
+		goto out;
 
 	if (!(opts.user_options & FT_OPT_SIZE)) {
 		for (i = 0; i < TEST_CNT; i++) {
@@ -508,37 +470,27 @@ static int run(void)
 		if (ret)
 			goto out;
 	}
-	
-	while (credits < max_credits)
-		poll_all_sends();
-	
+
+	/* Finalize before closing ep */
 	ft_finalize(ep, scq, rcq, remote_fi_addr);
 out:
 	free_ep_res();
-	ret = fi_close(&av->fid);
-	if (ret != 0) {
-		FT_PRINTERR("fi_close", ret);
-	}
-	ret = fi_close(&dom->fid);
-	if (ret != 0) {
-		FT_PRINTERR("fi_close", ret);
-	}
-	ret = fi_close(&fab->fid);
-	if (ret != 0) {
-		FT_PRINTERR("fi_close", ret);
-	}
-
+	fi_close(&dom->fid);
+	fi_close(&fab->fid);
 	return ret;
 }
 
 int main(int argc, char **argv)
 {
-	int ret, op;
+	int op, ret;
 	opts = INIT_OPTS;
+	opts.transfer_size = 64;
 
 	hints = fi_allocinfo();
-	if (!hints)
+	if (!hints) {
+		FT_PRINTERR("fi_allocinfo", -FI_ENOMEM);
 		return EXIT_FAILURE;
+	}
 
 	while ((op = getopt(argc, argv, "h" CS_OPTS INFO_OPTS)) != -1) {
 		switch (op) {
@@ -548,7 +500,7 @@ int main(int argc, char **argv)
 			break;
 		case '?':
 		case 'h':
-			ft_csusage(argv[0], "Ping pong client and server using UD.");
+			ft_csusage(argv[0], "Ping pong client and server using inject.");
 			return EXIT_FAILURE;
 		}
 	}
@@ -556,18 +508,18 @@ int main(int argc, char **argv)
 	if (optind < argc)
 		opts.dst_addr = argv[optind];
 
-	hints->ep_attr->type = FI_EP_DGRAM;
+	hints->ep_attr->type = FI_EP_RDM;
 	hints->caps = FI_MSG;
-	hints->mode = FI_LOCAL_MR | FI_MSG_PREFIX;
-	hints->addr_format = FI_SOCKADDR;
+	hints->mode = FI_CONTEXT;
 
-	if (opts.prhints) {
-		printf("%s", fi_tostr(hints, FI_TYPE_INFO));
-		ret = EXIT_SUCCESS;
-	} else {
-		ret = run();
-	}
+	if (opts.transfer_size)
+		hints->tx_attr->inject_size = opts.transfer_size;
+	else
+		hints->tx_attr->inject_size = 16;
+
+	ret = run();
+
 	fi_freeinfo(hints);
 	fi_freeinfo(fi);
-	return ret;
+	return -ret;
 }
