+++ pingpong/rdm_tagged_pingpong.c
@@ -39,15 +39,16 @@
 #include <rdma/fi_errno.h>
 #include <rdma/fi_endpoint.h>
 #include <rdma/fi_cm.h>
+#include <rdma/fi_tagged.h>
 
 #include "shared.h"
 
 static struct cs_opts opts;
-static int max_inject_size;
 static int max_credits = 128;
+static int credits = 128;
 static char test_name[10] = "custom";
 static struct timespec start, end;
-static void *send_buf, *recv_buf;
+static void *buf;
 static size_t buffer_size;
 
 static struct fi_info *fi, *hints;
@@ -55,31 +56,69 @@ static struct fi_info *fi, *hints;
 static struct fid_fabric *fab;
 static struct fid_domain *dom;
 static struct fid_ep *ep;
-static struct fid_cq *rcq;
-static struct fid_cq *scq;
+static struct fid_cq *rcq, *scq;
 static struct fid_av *av;
 static struct fid_mr *mr;
 static void *local_addr, *remote_addr;
 static size_t addrlen = 0;
 static fi_addr_t remote_fi_addr;
-struct fi_context fi_ctx_send;
+struct fi_context fi_ctx_tsend;
+struct fi_context fi_ctx_trecv;
 struct fi_context fi_ctx_recv;
 struct fi_context fi_ctx_av;
 
-static int send_xfer(int size)
+static uint64_t tag_data = 0;
+static uint64_t tag_control = 0x12345678;
+
+int wait_for_completion_tagged(struct fid_cq *cq, int num_completions)
 {
 	int ret;
+	struct fi_cq_tagged_entry comp;
 
-	ret = fi_inject(ep, send_buf, (size_t) size, remote_fi_addr);
+	while (num_completions > 0) {
+		ret = fi_cq_read(cq, &comp, 1);
+		if (ret > 0) {
+			num_completions--;
+		} else if (ret < 0 && ret != -FI_EAGAIN) {
+			FT_PRINTERR("fi_cq_read", ret);
+			return ret;
+		}
+	}
+	return 0;
+}
+
+static int send_xfer(int size)
+{
+	struct fi_cq_entry comp;
+	int ret;
+
+	while (!credits) {
+		ret = fi_cq_read(scq, &comp, 1);
+		if (ret > 0) {
+			goto post;
+		} else if (ret < 0 && ret != -FI_EAGAIN) {
+			if (ret == -FI_EAVAIL) {
+				cq_readerr(scq, "scq");
+			} else {
+				FT_PRINTERR("fi_cq_read", ret);
+			}
+			return ret;
+		}
+	}
+
+	credits--;
+post:
+	ret = fi_tsend(ep, buf, (size_t) size, fi_mr_desc(mr), remote_fi_addr,
+			tag_data, &fi_ctx_tsend);
 	if (ret)
-		FT_PRINTERR("fi_inject", ret);
+		FT_PRINTERR("fi_tsend", ret);
 
 	return ret;
 }
 
 static int recv_xfer(int size)
 {
-	struct fi_cq_entry comp;
+	struct fi_cq_tagged_entry comp;
 	int ret;
 
 	do {
@@ -94,10 +133,11 @@ static int recv_xfer(int size)
 		}
 	} while (ret == -FI_EAGAIN);
 
-	ret = fi_recv(ep, recv_buf, buffer_size, fi_mr_desc(mr), remote_fi_addr,
-			&fi_ctx_recv);
+	/* Posting recv for next send. Hence tag_data + 1 */
+	ret = fi_trecv(ep, buf, buffer_size, fi_mr_desc(mr), remote_fi_addr,
+			tag_data + 1, 0, &fi_ctx_trecv);
 	if (ret)
-		FT_PRINTERR("fi_recv", ret);
+		FT_PRINTERR("fi_trecv", ret);
 
 	return ret;
 }
@@ -106,11 +146,14 @@ static int send_msg(int size)
 {
 	int ret;
 
-	ret = fi_inject(ep, send_buf, (size_t) size, remote_fi_addr);
+	ret = fi_tsend(ep, buf, (size_t) size, fi_mr_desc(mr), remote_fi_addr,
+			tag_control, &fi_ctx_tsend);
 	if (ret) {
-		FT_PRINTERR("fi_inject", ret);
+		FT_PRINTERR("fi_tsend", ret);
 		return ret;
 	}
+
+	ret = wait_for_completion_tagged(scq, 1);
 
 	return ret;
 }
@@ -119,13 +162,14 @@ static int recv_msg(void)
 {
 	int ret;
 
-	ret = fi_recv(ep, recv_buf, buffer_size, fi_mr_desc(mr), 0, &fi_ctx_recv);
+	ret = fi_trecv(ep, buf, buffer_size, fi_mr_desc(mr), remote_fi_addr,
+		       tag_control, 0, &fi_ctx_trecv);
 	if (ret) {
-		FT_PRINTERR("fi_recv", ret);
+		FT_PRINTERR("fi_trecv", ret);
 		return ret;
 	}
 
-	ret = wait_for_completion(rcq, 1);
+	ret = wait_for_completion_tagged(rcq, 1);
 
 	return ret;
 }
@@ -134,19 +178,26 @@ static int sync_test(void)
 {
 	int ret;
 
+	ret = wait_for_completion_tagged(scq, max_credits - credits);
+	if (ret) {
+		return ret;
+	}
+	credits = max_credits;
+
 	ret = opts.dst_addr ? send_xfer(16) : recv_xfer(16);
 	if (ret)
 		return ret;
 
-	return opts.dst_addr ? recv_xfer(16) : send_xfer(16);
+	ret = opts.dst_addr ? recv_xfer(16) : send_xfer(16);
+
+	tag_data++;
+
+	return ret;
 }
 
 static int run_test(void)
 {
 	int ret, i;
-
-	if (opts.transfer_size > max_inject_size) 
-		return 0;
 
 	ret = sync_test();
 	if (ret)
@@ -163,6 +214,8 @@ static int run_test(void)
 				 send_xfer(opts.transfer_size);
 		if (ret)
 			goto out;
+
+		tag_data++;
 	}
 	clock_gettime(CLOCK_MONOTONIC, &end);
 
@@ -184,8 +237,7 @@ static void free_ep_res(void)
 	fi_close(&mr->fid);
 	fi_close(&rcq->fid);
 	fi_close(&scq->fid);
-	free(send_buf);
-	free(recv_buf);
+	free(buf);
 }
 
 static int alloc_ep_res(struct fi_info *fi)
@@ -196,9 +248,8 @@ static int alloc_ep_res(struct fi_info *fi)
 
 	buffer_size = opts.user_options & FT_OPT_SIZE ?
 			opts.transfer_size : test_size[TEST_CNT - 1].size;
-	send_buf = malloc(buffer_size);
-	recv_buf = malloc(buffer_size);
-	if (!send_buf || !recv_buf) {
+	buf = malloc(buffer_size);
+	if (!buf) {
 		perror("malloc");
 		return -1;
 	}
@@ -207,24 +258,21 @@ static int alloc_ep_res(struct fi_info *fi)
 	cq_attr.format = FI_CQ_FORMAT_CONTEXT;
 	cq_attr.wait_obj = FI_WAIT_NONE;
 	cq_attr.size = max_credits << 1;
-
-	ret = fi_cq_open(dom, &cq_attr, &rcq, NULL);
+	ret = fi_cq_open(dom, &cq_attr, &scq, NULL);
 	if (ret) {
 		FT_PRINTERR("fi_cq_open", ret);
 		goto err1;
 	}
 
-	ret = fi_cq_open(dom, &cq_attr, &scq, NULL);
+	ret = fi_cq_open(dom, &cq_attr, &rcq, NULL);
 	if (ret) {
 		FT_PRINTERR("fi_cq_open", ret);
 		goto err2;
 	}
 
-	/* Memory registration not required for send_buf since we use fi_inject.
-	 * fi_inject copies the buffer of data that needs to be sent. */
-	ret = fi_mr_reg(dom, recv_buf, buffer_size, 0, 0, 0, 0, &mr, NULL);
+	ret = fi_mr_reg(dom, buf, buffer_size, 0, 0, 0, 0, &mr, NULL);
 	if (ret) {
-		FT_PRINTERR("fi_cq_open", ret);
+		FT_PRINTERR("fi_mr_reg", ret);
 		goto err3;
 	}
 
@@ -253,12 +301,11 @@ err5:
 err4:
 	fi_close(&mr->fid);
 err3:
-	fi_close(&scq->fid);
-err2:
 	fi_close(&rcq->fid);
+err2:
+	fi_close(&scq->fid);
 err1:
-	free(send_buf);
-	free(recv_buf);
+	free(buf);
 	return ret;
 }
 
@@ -266,13 +313,13 @@ static int bind_ep_res(void)
 {
 	int ret;
 
-	ret = fi_ep_bind(ep, &rcq->fid, FI_RECV);
+	ret = fi_ep_bind(ep, &scq->fid, FI_SEND);
 	if (ret) {
 		FT_PRINTERR("fi_ep_bind", ret);
 		return ret;
 	}
 
-	ret = fi_ep_bind(ep, &scq->fid, FI_SEND);
+	ret = fi_ep_bind(ep, &rcq->fid, FI_RECV);
 	if (ret) {
 		FT_PRINTERR("fi_ep_bind", ret);
 		return ret;
@@ -290,11 +337,19 @@ static int bind_ep_res(void)
 		return ret;
 	}
 
+	/* Post the first recv buffer */
+	ret = fi_recv(ep, buf, buffer_size, fi_mr_desc(mr), 0, &fi_ctx_recv);
+	if (ret) {
+		FT_PRINTERR("fi_recv", ret);
+		return ret;
+	}
+
 	return ret;
 }
 
 static int init_fabric(void)
 {
+	struct fi_info *fi;
 	uint64_t flags = 0;
 	char *node, *service;
 	int ret;
@@ -315,15 +370,6 @@ static int init_fabric(void)
 	if (ret) {
 		FT_PRINTERR("fi_getinfo", ret);
 		return ret;
-	}
-	
-	/* check max msg size */
-	max_inject_size = fi->tx_attr->inject_size;
-	if ((opts.user_options & FT_OPT_SIZE) &&
-	    (opts.transfer_size > max_inject_size)) {
-		fprintf(stderr, "Msg size greater than max inject size\n");
-		ret = -FI_EINVAL;
-		goto err0;
 	}
 
 	/* We use provider MR attributes and direct address (no offsets) 
@@ -399,8 +445,8 @@ static int init_av(void)
 		}
 
 		/* Send local addr size and local addr */
-		memcpy(send_buf, &addrlen, sizeof(size_t));
-		memcpy(send_buf + sizeof(size_t), local_addr, addrlen);
+		memcpy(buf, &addrlen, sizeof(size_t));
+		memcpy(buf + sizeof(size_t), local_addr, addrlen);
 		ret = send_msg(sizeof(size_t) + addrlen);
 		if (ret)
 			return ret;
@@ -416,9 +462,9 @@ static int init_av(void)
 		if (ret)
 			return ret;
 
-		memcpy(&addrlen, recv_buf, sizeof(size_t));
+		memcpy(&addrlen, buf, sizeof(size_t));
 		remote_addr = malloc(addrlen);
-		memcpy(remote_addr, recv_buf + sizeof(size_t), addrlen);
+		memcpy(remote_addr, buf + sizeof(size_t), addrlen);
 
 		ret = fi_av_insert(av, remote_addr, 1, &remote_fi_addr, 0, 
 				&fi_ctx_av);
@@ -434,10 +480,10 @@ static int init_av(void)
 	}
 
 	/* Post first recv */
-	ret = fi_recv(ep, recv_buf, buffer_size, fi_mr_desc(mr), remote_fi_addr,
-			&fi_ctx_recv);
+	ret = fi_trecv(ep, buf, buffer_size, fi_mr_desc(mr), remote_fi_addr,
+			tag_data, 0, &fi_ctx_trecv);
 	if (ret)
-		FT_PRINTERR("fi_recv", ret);
+		FT_PRINTERR("fi_trecv", ret);
 
 	return ret;
 }
@@ -471,6 +517,7 @@ static int run(void)
 			goto out;
 	}
 
+	wait_for_completion_tagged(scq, max_credits - credits);
 	/* Finalize before closing ep */
 	ft_finalize(ep, scq, rcq, remote_fi_addr);
 out:
@@ -484,13 +531,10 @@ int main(int argc, char **argv)
 {
 	int op, ret;
 	opts = INIT_OPTS;
-	opts.transfer_size = 64;
 
 	hints = fi_allocinfo();
-	if (!hints) {
-		FT_PRINTERR("fi_allocinfo", -FI_ENOMEM);
+	if (!hints)
 		return EXIT_FAILURE;
-	}
 
 	while ((op = getopt(argc, argv, "h" CS_OPTS INFO_OPTS)) != -1) {
 		switch (op) {
@@ -500,7 +544,7 @@ int main(int argc, char **argv)
 			break;
 		case '?':
 		case 'h':
-			ft_csusage(argv[0], "Ping pong client and server using inject.");
+			ft_csusage(argv[0], "Ping pong client and server using tagged messages.");
 			return EXIT_FAILURE;
 		}
 	}
@@ -509,22 +553,12 @@ int main(int argc, char **argv)
 		opts.dst_addr = argv[optind];
 
 	hints->ep_attr->type = FI_EP_RDM;
-	hints->caps = FI_MSG;
+	hints->caps = FI_MSG | FI_TAGGED;
 	hints->mode = FI_CONTEXT;
 
-	if (opts.transfer_size)
-		hints->tx_attr->inject_size = opts.transfer_size;
-	else
-		hints->tx_attr->inject_size = 16;
-
-	if (opts.prhints) {
-		printf("%s", fi_tostr(&hints, FI_TYPE_INFO));
-		ret = EXIT_SUCCESS;
-	} else {
-		ret = run();
-	}
+	ret = run();
 
 	fi_freeinfo(hints);
 	fi_freeinfo(fi);
-	return ret;
+	return -ret;
 }
