+++ simple/rdm_tagged_search.c
@@ -520,5 +520,5 @@ int main(int argc, char **argv)
 	ret = run();
 	fi_freeinfo(hints);
 	fi_freeinfo(fi);
-	return ret;
+	return -ret;
 }
